requirements.txt
playwright=1.50.0
python-dotenv
scrapegraphai
pydantic
openai
tenacity